+++ packages/jsii-pacmak/test/expected.jsii-calc-base/java/src/main/java/software/amazon/jsii/tests/calculator/base/$Module.java
@@ -7,7 +7,7 @@ import software.amazon.jsii.JsiiModule;
 
 public final class $Module extends JsiiModule {
     public $Module() {
-        super("@scope/jsii-calc-base", "0.11.3", $Module.class, "jsii-calc-base@0.11.3.jsii.tgz");
+        super("@scope/jsii-calc-base", "0.12.0", $Module.class, "jsii-calc-base@0.12.0.jsii.tgz");
     }
 
     @Override