+++ packages/jsii-pacmak/test/expected.jsii-calc-lib/dotnet/Amazon.JSII.Tests.CalculatorPackageId.LibPackageId/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using Amazon.JSII.Runtime.Deputy;
 
-[assembly: JsiiAssembly("@scope/jsii-calc-lib", "0.11.3", "scope-jsii-calc-lib-0.11.3.tgz")]
+[assembly: JsiiAssembly("@scope/jsii-calc-lib", "0.12.0", "scope-jsii-calc-lib-0.12.0.tgz")]