+++ packages/jsii-pacmak/test/expected.jsii-calc/dotnet/Amazon.JSII.Tests.CalculatorPackageId/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using Amazon.JSII.Runtime.Deputy;
 
-[assembly: JsiiAssembly("jsii-calc", "0.11.3", "jsii-calc-0.11.3.tgz")]
+[assembly: JsiiAssembly("jsii-calc", "0.12.0", "jsii-calc-0.12.0.tgz")]